server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["fastapi"], version = "^0.184.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.16"}
asyncpg = "^0.27.0"
asyncio = "^3.4.3"
fastapi = "^0.97.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
mypy = "^1.3.0"
black = "^23.3.0"
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
